requirements.txt
python-dotenv
numpy
pandas
pymongo
pymongo
certifi
pymongo[srv]
scikit-learn
dill
scipy
PyYAML



##at last do it
## -e .